+++ app/src/main/java/com/example/treadmillassistant/backend/workout/WorkoutPhase.kt
@@ -1,10 +1,11 @@
 package com.example.treadmillassistant.backend.workout
 
-class WorkoutPhase(var duration: Int = 0,
-                   var speed: Double = 0.0,
-                   var tilt: Double = 0.0,
-                   var workoutPlanID: Int = -1,
-                   var orderNumber: Int = -1,
-                   val isFinished: Boolean = false,
-                   var ID: Long =-1) {
+class WorkoutPhase(
+    var duration: Int = 0,
+    var speed: Double = 0.0,
+    var tilt: Double = 0.0,
+    var workoutPlanID: Long = -1,
+    var orderNumber: Int = -1,
+    val isFinished: Boolean = false,
+    var ID: Long =-1) {
 }
